areas/xdp-tutorial.org add intro text
The XDP-tutorial is more than a tutorial. We want to evolve this to become a
resource for XDP/BPF developer to learn good programming practices. This
also means aspiring to high code quality, and show good practise for error
handling, and show how code are placed into common/ directory files/objects.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>