areas/mem: xdp_mem_track0* hint driver devel should use this
These scripts are not only useful when developing the kernel code
for mem_disconnect work-queue.

It is also relevant for NIC driver developer, as this can identify
bugs in drivers, if leaking page_pool pages.  This can fairly easily
happen in drivers error handling paths, when converting over a driver.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>