conference/NetDevConf2020: Slide: The page_pool allocator
Try to motivate why we introduce the page_pool memory allocator.

The hardest part is changing the driver memory model to be
compatible with XDP, which is why I created to page_pool,
to make this easier for drivers.

XDP itself is also about adding flexiblity to support
different RX memory models via xdp_mem_type, where page_pool
is just one of the supported models. But it is a little
bit out of scope for this slide.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>