Added a new ESPRESSObin setup howto for Fedora 35
I based this setup on Jesper's Fedora 31. It is functional as is. The
parts that I finished are getting a usable image and booting it. In a
later commit, I will brush up on the software after getting the image up
and running, which works pretty much from Jersper's documentation.

I have not tried the default kernel from the raw image because I
compiled my own, so I want to try that. And I am going to add a section
about zram and how to disable it if you don't compile support for it in
your kernel. Otherwise, it is complete.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>