areas/latency: limit what is monitored by script
The script softirq_net_latency_safe.bt attempts to be safe
to run on production systems with a lot of CPUs.

The scripting language bpftrace is a little too eager to
create BPF hash-maps that comes with overhead.

Also remove measuring the softirq runtime, but instead
focus on the latency from raise-to-run. As it was found
that softirq_exit tracepoint comes with strange overhead[1].

[1] https://github.com/cloudflare/ebpf_exporter/pull/300

Signed-off-by: Jesper Dangaard Brouer <jesper@cloudflare.com>