areas/latency: softirq_net_latency report 'comm'
For events over threshold also report current running process.

For "High IRQ-to-softirq latency" this is expected to be ksoftirqd.

When running this in production, we were surprised to see that on
occations 'nft' were being reported.  This is the nftables command
tool updating ruleset. 

More investigations are needed.  Initial investigations point at
kernel function nft_rbtree_insert and the BH disable/enabled
write_lock_bh() interaction.

Signed-off-by: Jesper Dangaard Brouer <hawk@kernel.org>