softirq_net_latency: use global histogram maps
This tool is for command line usage and on production there
is too many CPU to keep track of.  Making it difficult to
get something useful out of the many individual CPU
based historgrams.

This is of-cause not good for performance, but is it okay
to increase load on a single server in a large fleet and 
this is usually part on an experiment anyhow.

In production deployment we will use the more optimized
version of this script converted to restricted-C BPF code
in ebpf-exporter project.

Signed-off-by: Jesper Dangaard Brouer <hawk@kernel.org>